Update: Move template actions to the editor store.
