+++ packages/edit-site/src/components/page-templates-template-parts/index.js
@@ -42,11 +42,7 @@ import {
 	LAYOUT_TABLE,
 	LAYOUT_LIST,
 } from '../../utils/constants';
-import {
-	resetTemplateAction,
-	deleteTemplateAction,
-	renameTemplateAction,
-} from './actions';
+
 import usePatternSettings from '../page-patterns/use-pattern-settings';
 import { unlock } from '../../lock-unlock';
 import AddNewTemplatePart from './add-new-template-part';
@@ -200,6 +196,14 @@ function Preview( { item, viewType } ) {
 		</ExperimentalBlockEditorProvider>
 	);
 }
+
+const TEMPLATE_ACTIONS = [
+	'edit-post',
+	'reset-template',
+	'rename-template',
+	'view-post-revisions',
+	'delete-template',
+];
 
 export default function PageTemplatesTemplateParts( { postType } ) {
 	const { params } = useLocation();
@@ -361,20 +365,8 @@ export default function PageTemplatesTemplateParts( { postType } ) {
 		},
 		[ history ]
 	);
-	const [ editAction, viewRevisionsAction ] = usePostActions(
-		onActionPerformed,
-		[ 'edit-post', 'view-post-revisions' ]
-	);
-	const actions = useMemo(
-		() => [
-			editAction,
-			resetTemplateAction,
-			renameTemplateAction,
-			viewRevisionsAction,
-			deleteTemplateAction,
-		],
-		[ editAction, viewRevisionsAction ]
-	);
+
+	const actions = usePostActions( onActionPerformed, TEMPLATE_ACTIONS );
 
 	const onChangeView = useCallback(
 		( newView ) => {
